Use `vim-gitbranch` for Git branch in statusline
Use the `vim-gitbranch` plugin for getting the Git branch in the
statusline rather than a custom function using a call to `system()`. The
`system()` call happens on every (?) keystroke, meaning lots of calls,
which in turn lead to artifacts (stray characters, flickering of
cursor). The `vim-gitbranch` plugin is a light-weight solution that
fixes this issue.