Do not use bright colours with bold in Alacritty
The setting to use the _bright_ versions of colours automatically for
all bold statements was enabled, since I liked the colours it gave me in
Git. I didn't really understand this fully, though, and it gave
confusing results and oddly presenting colours in other tools (such as
experimentation with `jj`).

Disabling this setting allows for more fine-grained control over exactly
which colours should be used where, with the drawback that I had to
change the Git colours now, plus possibly changing more colours in the
future. This is fundamentally better, though, and will be less confusing
in the long run.