Replace volume monitor with at-execution script
Replace the volume monitor script with a script that is mapped to volume
controls instead. This prevents notifications from being sent even when
volume is not changed, which sometimes happened when an app accessed
the volume controls without changing anything. The new script only sends
notifications when actually changing the volume or toggling mute.